pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-project"
version = "0.1.0"
description = "Agente LangGraph con patrón act→observe→reason, RAG local, tools, memoria, CLI y API opcional."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Tu Nombre", email = "tu.email@example.com" }]
license = { text = "MIT" }

dependencies = [
  "langgraph>=0.2.30",
  "langchain>=0.3.7",
  "langchain-openai>=0.2.5",
  "langchain-community>=0.3.7",
  "faiss-cpu>=1.8.0.post1",
  "pypdf>=4.2.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "uvicorn>=0.30.6",
  "fastapi>=0.115.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.8",
  "black>=24.8.0",
  "pytest>=8.2.0",
]
api = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.6",
]

[project.scripts]
agent-cli = "src.agent.cli.main:main"

[tool.setuptools]
package-dir = {"" = ""}

[tool.setuptools.packages.find]
where = ["src"]
include = ["src.agent*"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E203"]

[tool.black]
line-length = 100
target-version = ["py310"]
